created refactored ttt.js with help argument and argument handler
